Initial commit.
Includes a base framework for a jQuery Gallery plugin.
HTML is currently static, but should later be generated by the jQuery plugin when loading data from flickr.
Features now is, fade in and out effects on thumbnails. Allow the uses to select an image and updates the viewer with the 'selected' image's url displaying a larger version of the image.